+++ ChangeLog.md
@@ -6,6 +6,8 @@
 - Add `RunInRenderThread` to support initialization of low level OpenGL resources.
 - Set correct `WidgetEnv` viewport from scroll (it looked good because of scissoring).
 - Add OpenGL example.
+- Fix issue with scrollbars using child coordinates for detecting clicks.
+- Add `containerDrawDecorations`. Simplify button/externalLink internals.
 
 ### 1.0.0.3